Onshape CSS Surgery to Load Static Hosted Fonts
This reduces server requests and makes the app load faster. Fonts required by Onshape are statically hosted, and the css references to them in the woolsthorpe css are found and replaced to attempt to load the statically hosted fonts before asking Onshape for them.
This is okay to do because the fonts should never change.