Created Userinfo, Nodestore, Appstorage classes
Appstorage is a generic class used to store a specified data type in an app elements. Nodestore extends Appstorage, allowing multiple Nodestore objects to be created. Userinfo also extends Appstorage, but it will only use the initAsSingleton (and initFromExisting) method of Appstorage, as there should only every be a single userinfo document (in Created By Me).